remove obsolte rasp_fs.h file and rename rasp_fs2.h | The log is now regenerated from persistent storage during boards initial phase | add RASP_DEBUG flag - used by the preprocessor in order to add debug code (serial prints)

+++ lib/ServerState/ServerState.cpp
@@ -49,7 +49,7 @@ Message * ServerState::dispatch(Message *msg) {
 // TODO: set persistent values
 Message * ServerState::handleRequestVoteReq(uint32_t term,
                                             uint32_t candidateID,
-                                            uint8_t  lastLogIndex,
+                                            uint16_t lastLogIndex,
                                             uint32_t lastLogTerm) {
     rvRes.term        = currentTerm;
     rvRes.voteGranted = false;
